+++ fpsim/locations/data_utils.py
@@ -11,6 +11,7 @@ from scipy import interpolate as si
 import starsim as ss
 from fpsim import defaults as fpd
 import fpsim.shared_data as sd
+from pathlib import Path
 
 sd_dir = os.path.dirname(sd.__file__)  # path to the shared_data directory
 
@@ -26,7 +27,7 @@ def read_data(location, filename, **kwargs):
     data_path = loc_mod.filenames()['base']
     try:
         # Read data from data_path
-        df = pd.read_csv(data_path / filename, **kwargs)
+        df = pd.read_csv(Path(data_path) / filename, **kwargs)
     except FileNotFoundError:
         # Try one level up; likely a regional location so pull from country data
         fallback_path = data_path.parent.parent / 'data'